+++ e2e/key/key.go
@@ -487,7 +487,7 @@ func (c *ctx) apptainerKeyRemoveOpts(t *testing.T) {
 			name:    "remove private should succeed",
 			command: "key remove",
 			profile: e2e.UserProfile,
-			args:    []string{"--private", keyMap["key1"]},
+			args:    []string{"--secret", keyMap["key1"]},
 			exit:    0,
 		},
 		{
@@ -526,7 +526,7 @@ func (c *ctx) apptainerKeyRemoveOpts(t *testing.T) {
 			name:               "remove private should fail because key does not exist in keyring",
 			command:            "key remove",
 			profile:            e2e.UserProfile,
-			args:               []string{"--private", keyMap["key1"]},
+			args:               []string{"--secret", keyMap["key1"]},
 			expectedErrorRegex: "FATAL:",
 			exit:               255,
 		},
@@ -782,7 +782,7 @@ func E2ETests(env e2e.TestEnv) testhelper.Tests {
 		"ordered": func(t *testing.T) {
 			t.Run("keyCmd", c.apptainerKeyCmd)                       // Run all the tests in order
 			t.Run("keyNewpairWithLen", c.apptainerKeyNewpairWithLen) // We run a separate test for `key newpair --bit-length` because it requires handling a keyring a specific way
-			t.Run("keyRemoveOpts", c.apptainerKeyRemoveOpts)         // run a separated test for `key remove --public/--private/--both`
+			t.Run("keyRemoveOpts", c.apptainerKeyRemoveOpts)         // run a separated test for `key remove --public/--secret/--both`
 		},
 	}
 }